backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
openai>=1.0.0
python-dotenv==1.0.1
